Fix data type conversion from np.int64 to int in make_template_full kernel

+++ dsl/templates.py
@@ -194,7 +194,7 @@ def make_template_full(coord, parallel_batch: bool):
         def kernel(imgs, out_h_arr, out_w_arr, patch_offsets, mean, std, scale,
                    output, patch_size, temporal_patch_size, merge_size):
             for b in prange(len(imgs)):
-                img = imgs[np.int64(b)]
+                img = imgs[int(b)]
                 h = img.shape[0]
                 w = img.shape[1]
                 c = img.shape[2]
@@ -227,7 +227,7 @@ def make_template_full(coord, parallel_batch: bool):
         def kernel(imgs, out_h_arr, out_w_arr, patch_offsets, mean, std, scale,
                    output, patch_size, temporal_patch_size, merge_size):
             for b in range(len(imgs)):
-                img = imgs[np.int64(b)]
+                img = imgs[int(b)]
                 h = img.shape[0]
                 w = img.shape[1]
                 c = img.shape[2]
